Move app initialization to a separate class to reuse the logic in tests
